Create a 'copy' button to automatically copy the signature element.
Turns out, we didn't need the iframe. Also, it's hard to access the contents
of the iframe from outside of it, which the 'copy' button needs, so the iframe
is gone.

react-clipboard.js is a wrapper for clipboard.js. It appears to work on mobile
safari. The other clipboard libaries I saw relied on Flash.
One issue is it doesn't look like I can supply custom elements for the copy
button. It only takes a string. The default button is ugly, so I used an anchor
element for the 'button'.

It's a small convenience library and we could probably modify it to take a
custom element.